ConstData: 	- removed asteroids for testing
FrameData:
	- changed interpolation settings

ProcessLazer:
	- moved lazer index handling to ship weapon
	- moved lazers for peer into its own static funciton
	- fixed error with delta time not factoring into movement speed

ProcessShipHealth:
	- forced health to zero when dead

ProcessShipMovement:
	- fixed error not factoring in DelatTime into movement

ProcessShipWeapons :
	- fixed error not incrementing lazer correctly

SimPrecessorSettingsInterface:
	- fixed error with lazer size beeing set the same as speed

ActiveGameManagerSceneTester:
	- added debugs for ship rotation and weapon cooldown
	- added game state view spawner
	- added debug out for chain start candidate
	- added debug out when connecting as additional peer

TestingSimManager:
	- fixed bug on startup requesting data at tick 0 instead of start
	of SimStateBuffer

GlobalMessagingState:
	- added new function for Assigned channels vs active channesl
	- changed  active channel cout to only count a channel if message
	has been recieved from it

NetworkGlobalMessageProcessor:
	- added debug out for percent of states recieved

WebRTCTransmitter:
	- updated code to work with the new interface for the
	WebrRTC Package

GameStateViewSpawner:
	- first pass at component that spawns ecs entities to represent
	sim game objects

FrameDataInterpolatorGen:
	- added extra code to correctly interpolate angles

FrameDataInterpolationTypeAttribute:
	- added circulat interpolation mode

+++ Assets/Code/ProjectSimDataInterpolation/InterpolatedFrameDataGen.cs
@@ -11,11 +11,11 @@ namespace SimDataInterpolation
 	{
 		public System.Int64[] m_lPeersAssignedToSlot;
 		public System.Int32[] m_bInput;
-		public System.Byte[] m_fixShipHealth;
+		public System.Single[] m_fixShipHealth;
 																																									
 		public System.Single[] m_fixShipHealthErrorOffset; 
 
-		public System.Byte[] m_fixShipHealthErrorAdjusted;
+		public System.Single[] m_fixShipHealthErrorAdjusted;
 
 		public System.Single[] m_fixShipHealDelayTimeOut;
 																																									
@@ -118,12 +118,12 @@ namespace SimDataInterpolation
 			
 			if(m_fixShipHealth == null || m_fixShipHealth.Length != fdaFrameData.m_fixShipHealth.Length)
 			{
-				m_fixShipHealth = new System.Byte[fdaFrameData.m_fixShipHealth.Length] ;
+				m_fixShipHealth = new System.Single[fdaFrameData.m_fixShipHealth.Length] ;
 
 																																									
 				m_fixShipHealthErrorOffset = new System.Single[fdaFrameData.m_fixShipHealth.Length] ;
 
-				m_fixShipHealthErrorAdjusted = new System.Byte[fdaFrameData.m_fixShipHealth.Length] ;
+				m_fixShipHealthErrorAdjusted = new System.Single[fdaFrameData.m_fixShipHealth.Length] ;
 			
 			}
